moved patch for GameLocationCharacterManager.RevealCharacter from GameLocationPositioningManagerPatcher to GameLocationCharacterManagerPatcher
